src/adk/pyproject.toml
[project]
name = "adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.70",
    "fastapi==0.118.3",
    "google-adk",
    "litellm>=1.79.3",
    "pydantic>=2.11.1,<3",
]

[dependency-groups]
dev = [
    "ipython>=9.8.0",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pudb>=2025.1.3",
    "seaborn>=0.13.2",
]

[tool.uv.sources]
google-adk = { git = "https://github.com/mmphego/adk-python", rev = "fix/pydantic-v2-swagger-ui" }
